Use a manual bash script (rather than logrotate) to rotate .bash_history
The logic used by logrotate is a bit different from what we'd want to
have when rotating .bash_history -- namely, the fact that we don't want
to start from an empty history file after the rotation; a set of N
entries should be kept. So we've replaced the logrotate config with a
simple bash script.