Add a vim map for python files to fold typical multi-line statements
It is common for other people to break "no-logic" lines into multiple
lines, which I find a bit anoying as that often "polutes" a snippet of
code with several lines that no one cares about 95% of the time. So this
vim map attempts to do something similar to what the JetBrain IDEs do --
it "folds" the multiple lines into a single one, half-hidding their
content until the user explictly opens the fold.